+++ rewarped/warp_utils.py
@@ -24,9 +24,8 @@ def sim_eval_kinematic_fk(model, state_in, state_out, sim_dt, sim_substeps, cont
     wp.sim.eval_fk(model, state_out.joint_q, state_out.joint_qd, None, state_out)
 
 
-def sim_update(update_params, sim_params, states, control):
-    tape, integrator, model, use_graph_capture, synchronize = update_params
-    sim_substeps, sim_dt, eval_kinematic_fk, eval_ik = sim_params
+def sim_update(sim_params, model, states, control):
+    integrator, sim_substeps, sim_dt, eval_kinematic_fk, eval_ik = sim_params
     state_in, states_mid, state_out = states
 
     state_0 = state_in
@@ -48,8 +47,9 @@ def sim_update(update_params, sim_params, states, control):
         wp.sim.eval_ik(model, state_out, state_out.joint_q, state_out.joint_qd)
 
 
-def sim_update_inplace(model, integrator, state_0, state_1, sim_dt, sim_substeps, control, sim_params):
-    eval_kinematic_fk, eval_ik = sim_params
+def sim_update_inplace(sim_params, model, states, control):
+    integrator, sim_substeps, sim_dt, eval_kinematic_fk, eval_ik = sim_params
+    state_0, state_1 = states
 
     for i in range(sim_substeps):
         if eval_kinematic_fk:
@@ -63,4 +63,4 @@ def sim_update_inplace(model, integrator, state_0, state_1, sim_dt, sim_substeps
     if eval_ik:
         wp.sim.eval_ik(model, state_0, state_0.joint_q, state_0.joint_qd)
 
-    return state_0, state_1
+    return (state_0, state_1)