+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased -
+- Add `autograd.WarpKernelsFunction` for arbitrary warp kernels (single call)
+- Add model to `autograd.UpdateFunction` args
 - Change `Environment.render()` to use `state or self.state_0`
 - Rename `kinematic_fk` to `eval_kinematic_fk`
 - Add edge_count, tri_count, spring_count to `Environment.print_model_info()`